split up query utils file
