+++ posthog/hogql_queries/experiments/funnel_query_utils.py
@@ -1,0 +1,59 @@
+from posthog.hogql import ast
+from posthog.hogql.parser import parse_expr
+from posthog.models.team.team import Team
+from posthog.hogql_queries.experiments.base_query_utils import (
+    event_or_action_to_filter,
+)
+from posthog.schema import (
+    ActionsNode,
+    EventsNode,
+    ExperimentFunnelMetric,
+)
+
+
+def funnel_steps_to_filter(team: Team, funnel_steps: list[EventsNode | ActionsNode]) -> ast.Expr:
+    """
+    Returns the OR expression for a list of funnel steps. Will match if any of the funnel steps are true.
+    """
+    return ast.Or(exprs=[event_or_action_to_filter(team, funnel_step) for funnel_step in funnel_steps])
+
+
+def funnel_steps_to_window_funnel_expr(funnel_metric: ExperimentFunnelMetric) -> ast.Expr:
+    """
+    Returns the expression for the window funnel. The expression returns 1 if the user completed the whole funnel, 0 if they didn't.
+    """
+
+    funnel_steps_str = ", ".join([f"funnel_step = 'step_{i}'" for i, _ in enumerate(funnel_metric.series)])
+
+    # TODO: get conversion time window from funnel config
+    num_steps = len(funnel_metric.series)
+    conversion_time_window = 6048000000000000
+    return parse_expr(
+        f"windowFunnel({conversion_time_window})(toDateTime(timestamp), {funnel_steps_str}) = {num_steps}",
+        placeholders={
+            "conversion_time_window": ast.Constant(value=conversion_time_window),
+            "num_steps": ast.Constant(value=num_steps),
+        },
+    )
+
+
+def get_funnel_step_level_expr(team: Team, funnel_metric: ExperimentFunnelMetric) -> ast.Expr:
+    """
+    Returns the expression to get the funnel step level.
+
+    We reuse the filters that are being used to select events/actions in the funnel metric query,
+    and pass them into multiIf to get the funnel step level.
+    """
+
+    # Contains tuples of (filter: ast.Expr, step_name: ast.Constant)
+    filters_and_steps = [
+        (event_or_action_to_filter(team, funnel_step), ast.Constant(value=f"step_{i}"))
+        for i, funnel_step in enumerate(funnel_metric.series)
+    ]
+    # Flatten the list of tuples into a list of expressions to pass to multiIf
+    multi_if_args = [item for filter_value_pair in filters_and_steps for item in filter_value_pair]
+
+    # Last argument to multiIf is the default value
+    multi_if_args.append(ast.Constant(value="step_unknown"))
+
+    return ast.Call(name="multiIf", args=multi_if_args)
